feat:update and delete contact with auth user

+++ controllers/contactController.js
@@ -6,8 +6,8 @@ const Contact = require("../model/contactModel")
 //@access public
 
 const getContacts = asyncHandler( async(req, res) =>{
-    const contact= await Contact.find()
-    res.status(200).json(contact);
+    const contacts= await Contact.find({user_id:req.user.id})
+    res.status(200).json(contacts);
 })
 
 
@@ -17,12 +17,13 @@ const getContacts = asyncHandler( async(req, res) =>{
 const createContact =  asyncHandler(  async(req , res)=>{
     console.log('requested body',req.body)
     const {name ,email,phone}=req.body
-    if(!name||!email||!phone){
+    if(!name||!email||!phone){ 
         res.status(400)
         throw new Error('all Fields are required')
 
     }
-    const contact=await Contact.create({name,email,phone})
+   
+    const contact=await Contact.create({name,email,phone,user_id:req.user.id})
     res.status(201).json(contact)
 })
 
@@ -46,6 +47,10 @@ const updateContact=asyncHandler(async(req , res)=>{
     res.status(404)
     throw new Error("Contact not found")
  }
+if(contact.user_id!==req.user.id){
+    res.status(403)
+    throw new Error("user don't have permission to update other user contacts")
+}
  const updatedContact = await Contact.findByIdAndUpdate(req.params.id,req.body,{new:true})
     res.status(200).json({message:'updated Successfully'})
 })
@@ -57,6 +62,10 @@ const deleteContact =asyncHandler(async(req, res)=>{
     if(!contact){
         res.status(404)
         throw new Error('contact not Found')
+    }
+    if(contact.user_id.toString()!==req.user.id){
+        res.status(403)
+        throw new Error("user don't have permission to delete other user contacts")
     }
     const deleteContact= await Contact.findByIdAndDelete(req.params.id)
     res.status(200).json({message:` Delete contact for ${req.params.id}`,deleteContact:deleteContact})
